+++ lib/hanami/cli/commands/app/db/command.rb
@@ -31,7 +31,7 @@ module Hanami
             def run_command(klass, ...)
               klass.new(
                 out: out,
-                inflector: fs,
+                inflector: inflector,
                 fs: fs,
                 system_call: system_call,
               ).call(...)
@@ -39,59 +39,87 @@ module Hanami
 
             private
 
-            def databases(app: false, slice: nil)
-              if app
-                [database_for_app]
-              elsif slice
-                [database_for_slice(slice)]
-              else
-                all_databases
+            def databases(app: false, slice: nil, gateway: nil)
+              if gateway && !app && !slice
+                err.puts "When specifying --gateway, an --app or --slice must also be given"
+                exit 1
               end
+
+              databases =
+                if slice
+                  [database_for_slice(slice, gateway: gateway)]
+                elsif app
+                  [database_for_slice(self.app, gateway: gateway)]
+                else
+                  all_databases
+                end
+
+              databases.flatten
             end
 
-            def database_for_app
-              build_database(app)
-            end
-
-            def database_for_slice(slice)
+            def database_for_slice(slice, gateway: nil)
               unless slice.is_a?(Class) && slice < Hanami::Slice
                 slice_name = inflector.underscore(Shellwords.shellescape(slice)).to_sym
                 slice = app.slices[slice_name]
               end
 
-              build_database(slice)
+              ensure_database_slice slice
+
+              databases = build_databases(slice)
+
+              if gateway
+                databases.fetch(gateway.to_sym) do
+                  err.puts %(No gateway "#{gateway}" in #{slice})
+                  exit 1
+                end
+              else
+                databases.values
+              end
             end
 
-            def all_databases
+            def all_databases # rubocop:disable Metrics/AbcSize, Metrics/PerceivedComplexity
               slices = [app] + app.slices.with_nested
 
-              slices_by_database_url = slices.each_with_object({}) { |slice, hsh|
-                provider = slice.container.providers[:db]
-                next unless provider
+              slice_gateways_by_database_url = slices.each_with_object({}) { |slice, hsh|
+                db_provider_source = slice.container.providers[:db]&.source
+                next unless db_provider_source
 
-                database_url = provider.source.database_url
-                hsh[database_url] ||= []
-                hsh[database_url] << slice
+                db_provider_source.database_urls.each do |gateway, url|
+                  hsh[url] ||= []
+                  hsh[url] << {slice: slice, gateway: gateway}
+                end
               }
 
-              databases = slices_by_database_url.each_with_object([]) { |(url, slices), arr|
-                slices_with_config = slices.select { _1.root.join("config", "db").directory? }
+              slice_gateways_by_database_url.each_with_object([]) { |(url, slice_gateways), arr|
+                slice_gateways_with_config = slice_gateways.select {
+                  _1[:slice].root.join("config", "db").directory?
+                }
 
-                database = build_database(slices_with_config.first || slices.first)
+                db_slice_gateway = slice_gateways_with_config.first || slice_gateways.first
+                database = Utils::Database.database_class(url).new(
+                  slice: db_slice_gateway.fetch(:slice),
+                  gateway_name: db_slice_gateway.fetch(:gateway),
+                  system_call: system_call
+                )
 
-                warn_on_misconfigured_database database, slices_with_config
+                warn_on_misconfigured_database database, slice_gateways.map { _1.fetch(:slice) }
 
                 arr << database
               }
-
-              databases
             end
 
-            def build_database(slice)
-              Utils::Database[slice, system_call: system_call]
+            def build_databases(slice)
+              Utils::Database.from_slice(slice: slice, system_call: system_call)
             end
 
-            def warn_on_misconfigured_database(database, slices)
+            def ensure_database_slice(slice)
+              return if slice.container.providers[:db]
+
+              out.puts "#{slice} does not have a :db provider."
+              exit 1
+            end
+
+            def warn_on_misconfigured_database(database, slices) # rubocop:disable Metrics/AbcSize
               if slices.length > 1
                 out.puts <<~STR
                   WARNING: Database #{database.name} is configured for multiple config/db/ directories:
@@ -101,8 +129,11 @@ module Hanami
                   Migrating database using #{database.slice.slice_name.to_s.inspect} slice only.
 
                 STR
-              elsif slices.length < 1
-                relative_path = database.slice.root.relative_path_from(database.slice.app.root).join("config", "db").to_s
+              elsif !database.db_config_dir?
+                relative_path = database.slice.root
+                  .relative_path_from(database.slice.app.root)
+                  .join("config", "db").to_s
+
                 out.puts <<~STR
                   WARNING: Database #{database.name} expects the folder #{relative_path}/ to exist but it does not.