+++ spec/unit/hanami/cli/commands/app/db/create_spec.rb
@@ -49,26 +49,26 @@ RSpec.describe Hanami::CLI::Commands::App::DB::Create, :app_integration do
 
     describe "sqlite" do
       before do
-        ENV["DATABASE_URL"] = "sqlite://db/bookshelf_development.sqlite3"
+        ENV["DATABASE_URL"] = "sqlite://db/app.sqlite3"
       end
 
       it "creates the database" do
         command.call
 
-        expect(Hanami.app.root.join("db", "bookshelf_development.sqlite3").exist?).to be true
+        expect(Hanami.app.root.join("db", "app.sqlite3").exist?).to be true
 
         expect { Hanami.app["db.gateway"] }.not_to raise_error
 
-        expect(output).to include "database db/bookshelf_development.sqlite3 created"
+        expect(output).to include "database db/app.sqlite3 created"
       end
 
       it "does not create the database if it already exists" do
         FileUtils.mkdir(@dir.join("db"))
-        FileUtils.touch(@dir.join("db", "bookshelf_development.sqlite3"))
+        FileUtils.touch(@dir.join("db", "app.sqlite3"))
 
         command.call
 
-        expect(output).to include "database db/bookshelf_development.sqlite3 created"
+        expect(output).to include "database db/app.sqlite3 created"
       end
     end
 
@@ -133,67 +133,121 @@ RSpec.describe Hanami::CLI::Commands::App::DB::Create, :app_integration do
 
     describe "sqlite" do
       before do
-        ENV["DATABASE_URL"] = "sqlite://db/bookshelf_development.sqlite3"
-        ENV["MAIN__DATABASE_URL"] = "sqlite://db/bookshelf_main_development.sqlite3"
+        ENV["DATABASE_URL"] = "sqlite://db/app.sqlite3"
+        ENV["MAIN__DATABASE_URL"] = "sqlite://db/main.sqlite3"
       end
 
       it "creates each database" do
         command.call
 
-        expect(Hanami.app.root.join("db", "bookshelf_development.sqlite3").exist?).to be true
-        expect(Hanami.app.root.join("db", "bookshelf_main_development.sqlite3").exist?).to be true
+        expect(Hanami.app.root.join("db", "app.sqlite3").exist?).to be true
+        expect(Hanami.app.root.join("db", "main.sqlite3").exist?).to be true
 
         expect { Hanami.app["db.gateway"] }.not_to raise_error
         expect { Main::Slice["db.gateway"] }.not_to raise_error
 
-        expect(output).to include "database db/bookshelf_development.sqlite3 created"
-        expect(output).to include "database db/bookshelf_main_development.sqlite3 created"
+        expect(output).to include "database db/app.sqlite3 created"
+        expect(output).to include "database db/main.sqlite3 created"
       end
 
       it "creates the app database when given --app" do
         command.call(app: true)
 
-        expect(Hanami.app.root.join("db", "bookshelf_development.sqlite3").exist?).to be true
-        expect(Hanami.app.root.join("db", "bookshelf_main_development.sqlite3").exist?).to be false
+        expect(Hanami.app.root.join("db", "app.sqlite3").exist?).to be true
+        expect(Hanami.app.root.join("db", "main.sqlite3").exist?).to be false
 
         expect { Hanami.app["db.gateway"] }.not_to raise_error
 
-        expect(output).to include "database db/bookshelf_development.sqlite3 created"
-        expect(output).not_to include "db/bookshelf_main_development.sqlite3"
+        expect(output).to include "database db/app.sqlite3 created"
+        expect(output).not_to include "db/main.sqlite3"
       end
 
       it "creates a slice database when given --slice" do
         command.call(slice: "main")
 
-        expect(Hanami.app.root.join("db", "bookshelf_main_development.sqlite3").exist?).to be true
-        expect(Hanami.app.root.join("db", "bookshelf_development.sqlite3").exist?).to be false
+        expect(Hanami.app.root.join("db", "main.sqlite3").exist?).to be true
+        expect(Hanami.app.root.join("db", "app.sqlite3").exist?).to be false
 
         expect { Main::Slice["db.gateway"] }.not_to raise_error
 
-        expect(output).to include "database db/bookshelf_main_development.sqlite3 created"
-        expect(output).not_to include "db/bookshelf_development.sqlite3"
+        expect(output).to include "database db/main.sqlite3 created"
+        expect(output).not_to include "db/app.sqlite3"
       end
 
       it "prints errors for any create commands that fail and exits with non-zero status" do
         allow(system_call).to receive(:call).and_call_original
         allow(system_call)
           .to receive(:call)
-          .with(a_string_matching(/sqlite3.+bookshelf_development.sqlite3/))
+          .with(a_string_matching(/sqlite3.+app.sqlite3/))
           .and_return Hanami::CLI::SystemCall::Result.new(exit_code: 2, out: "", err: "app-db-err")
 
         command.call
 
         expect { Main::Slice["db.gateway"] }.not_to raise_error
 
-        expect(Hanami.app.root.join("db", "bookshelf_development.sqlite3").exist?).to be false
-        expect(Hanami.app.root.join("db", "bookshelf_main_development.sqlite3").exist?).to be true
+        expect(Hanami.app.root.join("db", "app.sqlite3").exist?).to be false
+        expect(Hanami.app.root.join("db", "main.sqlite3").exist?).to be true
 
-        expect(output).to include "failed to create database db/bookshelf_development.sqlite3"
+        expect(output).to include "failed to create database db/app.sqlite3"
         expect(output).to include "app-db-err"
 
-        expect(output).to include "database db/bookshelf_main_development.sqlite3 created"
+        expect(output).to include "database db/main.sqlite3 created"
 
         expect(command).to have_received(:exit).with(2).once
+      end
+
+      context "app with gateways" do
+        def before_prepare
+          write "config/db/.keep", ""
+          ENV["DATABASE_URL__EXTRA"] = "sqlite://db/app_extra.sqlite3"
+        end
+
+        it "creates the databases for all the app's gateways when given --app" do
+          expect { command.call(app: true) }
+            .to change { Hanami.app.root.join("db", "app.sqlite3").exist? }.to(true)
+            .and change { Hanami.app.root.join("db", "app_extra.sqlite3").exist? }.to(true)
+
+          expect(output).to include_in_order(
+            "database db/app.sqlite3 created",
+            "database db/app_extra.sqlite3 created",
+          )
+        end
+
+        it "creates the database for an app's gateway when given --app and --gateway" do
+          expect { command.call(app: true, gateway: "extra") }
+            .to change { Hanami.app.root.join("db", "app_extra.sqlite3").exist? }.to(true)
+            .and not_change { Hanami.app.root.join("db", "app.sqlite3").exist? }.from(false)
+
+          expect(output).to include "database db/app_extra.sqlite3 created"
+          expect(output).not_to include "database/app.sqlite3"
+        end
+      end
+
+      context "slice with gateways" do
+        def before_prepare
+          write "slices/main/config/db/.keep", ""
+          ENV["MAIN__DATABASE_URL__EXTRA"] = "sqlite://db/main_extra.sqlite3"
+        end
+
+        it "creates the databases for all the slices's gateways when given --slice" do
+          expect { command.call(slice: "main") }
+            .to change { Hanami.app.root.join("db", "main.sqlite3").exist? }.to(true)
+            .and change { Hanami.app.root.join("db", "main_extra.sqlite3").exist? }.to(true)
+
+          expect(output).to include_in_order(
+            "database db/main.sqlite3 created",
+            "database db/main_extra.sqlite3 created",
+          )
+        end
+
+        it "creates the database for an app's gateway when given --app and --gateway" do
+          expect { command.call(slice: "main", gateway: "extra") }
+            .to change { Hanami.app.root.join("db", "main_extra.sqlite3").exist? }.to(true)
+            .and not_change { Hanami.app.root.join("db", "main.sqlite3").exist? }.from(false)
+
+          expect(output).to include "database db/main_extra.sqlite3 created"
+          expect(output).not_to include "database/main.sqlite3"
+        end
       end
     end